requirements.txt
selenium>=4.0.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=1.0.0